pyproject.toml
[project]
name = "alx-backend-python"
version = "0.1.0"
description = ""
authors = [
    {name = "Nii-Armah",email = "stanleyniiarmah7@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.4,<3.0.0)",
    "parameterized (>=0.9.0,<0.10.0)",
    "django (>=5.2.4,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "drf-nested-routers (>=0.94.2,<0.95.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-filter (>=25.1,<26.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
